+++ R/fit_prcmlpmm.R
@@ -84,13 +84,13 @@
 #' do.bootstrap = FALSE
 #' # IMPORTANT: set do.bootstrap = TRUE to compute the optimism correction!
 #' n.boots = ifelse(do.bootstrap, 100, 0)
-#' parallelize = TRUE
-#' # IMPORTANT: set parallelize = TRUE to speed computations up!
-#' if (!parallelize) n.cores = 1
-#' if (parallelize) {
+#' more.cores = FALSE
+#' # IMPORTANT: set more.cores = TRUE to speed computations up!
+#' if (!more.cores) n.cores = 2
+#' if (more.cores) {
 #'    # identify number of available cores on your machine
 #'    n.cores = parallel::detectCores()
-#'    if (is.na(n.cores)) n.cores = 1
+#'    if (is.na(n.cores)) n.cores = 2
 #' }
 #' 
 #' # step 1 of PRC-MLPMM: estimate the MLPMMs