+++ R/performance_prc.R
@@ -54,13 +54,13 @@
 # load fitted model
 #' data(fitted_prclmm)
 #' 
-#' parallelize = FALSE
-#' # IMPORTANT: set parallelize = TRUE to speed computations up!
-#' if (!parallelize) n.cores = 1
-#' if (parallelize) {
+#' more.cores = FALSE
+#' # IMPORTANT: set more.cores = TRUE to speed computations up!
+#' if (!more.cores) n.cores = 2
+#' if (more.cores) {
 #'    # identify number of available cores on your machine
 #'    n.cores = parallel::detectCores()
-#'    if (is.na(n.cores)) n.cores = 1
+#'    if (is.na(n.cores)) n.cores = 2
 #' }
 #'                    
 #' # compute the performance measures